Move util/args.rs into cli.rs
