+++ noita-proxy/src/util/mod.rs
@@ -1,3 +1,2 @@
-pub mod args;
 pub mod lang;
 pub mod steam_helper;
